UDF programmer
With this tool you can program an UDF on AccessB hardware.

The UDF must be created from the AccessB mplab project writting in the
UDF.C file and then compile the entire project, UDF programmer will
extract the UDF opcode from the file and will program it on the PIC of
the board.